+++ src/AppserverIo/Appserver/Core/Interfaces/SystemConfigurationInterface.php
@@ -136,4 +136,29 @@ interface SystemConfigurationInterface
      * @return string The username
      */
     public function getUser();
+
+    /**
+     * Array with the handler params to use.
+     *
+     * @return array
+     */
+    public function getParams();
+
+    /**
+     * Returns the param with the passed name casted to
+     * the specified type.
+     *
+     * @param string $name The name of the param to be returned
+     *
+     * @return mixed The requested param casted to the specified type
+     */
+    public function getParam($name);
+
+    /**
+     * Returns the params casted to the defined type
+     * as associative array.
+     *
+     * @return array The array with the casted params
+     */
+    public function getParamsAsArray();
 }